requirements.txt
requests>=2.33.0
python-dotenv>=1.2.0
pandas>=2.0.0
matplotlib>=3.7.0
rich>=13.0.0
